Rename `Stub::create()` to `Double::instance()` and `Stub::classname()` to `Double::classname()` (BC Break).
