+++ spec/Fixture/Plugin/Double/AbstractDoz.php
@@ -1,5 +1,5 @@
 <?php
-namespace Kahlan\Spec\Fixture\Plugin\Stub;
+namespace Kahlan\Spec\Fixture\Plugin\Double;
 
 abstract class AbstractDoz
 {
