+++ spec/Suite/Legacy/SpecificationSpec.php
@@ -6,7 +6,7 @@ use Exception;
 use Kahlan\Suite;
 use Kahlan\Specification;
 use Kahlan\Matcher;
-use Kahlan\Plugin\Stub;
+use Kahlan\Plugin\Double;
 
 describe("Specification", function() {
 
@@ -89,7 +89,7 @@ describe("Specification", function() {
                 $this->spec = new Specification([
                     'message' => 'runs a spec',
                     'closure' => function() {
-                        $stub = Stub::create();
+                        $stub = Double::instance();
                         $this->expect($stub)->toReceive('methodName');
                         $stub->methodName();
                     }
@@ -141,7 +141,7 @@ describe("Specification", function() {
 
                 $this->spec = new Specification([
                     'closure' => function() {
-                        $this->expect(Stub::create())->not->toReceive('helloWorld');
+                        $this->expect(Double::instance())->not->toReceive('helloWorld');
                     }
                 ]);
 
@@ -162,7 +162,7 @@ describe("Specification", function() {
 
                 $this->spec = new Specification([
                     'closure' => function() {
-                        $stub = Stub::create();
+                        $stub = Double::instance();
                         $this->expect($stub)->not->toReceive('methodName');
                     }
                 ]);
@@ -225,7 +225,7 @@ describe("Specification", function() {
                 $this->spec = new Specification([
                     'message' => 'runs a spec',
                     'closure' => function() {
-                        $stub = Stub::create();
+                        $stub = Double::instance();
                         $this->expect($stub)->toReceive('methodName');
                     }
                 ]);
@@ -276,7 +276,7 @@ describe("Specification", function() {
 
                 $this->spec = new Specification([
                     'closure' => function() {
-                        $stub = Stub::create();
+                        $stub = Double::instance();
                         $this->expect($stub)->not->toReceive('methodName');
                         $stub->methodName();
                     }
