+++ spec/Suite/SuiteSpec.php
@@ -11,7 +11,7 @@ use Kahlan\Suite;
 use Kahlan\Matcher;
 use Kahlan\Reporters;
 use Kahlan\Arg;
-use Kahlan\Plugin\Stub;
+use Kahlan\Plugin\Double;
 
 describe("Suite", function() {
 
@@ -814,7 +814,7 @@ describe("Suite", function() {
 
             });
 
-            $reporters = Stub::create();
+            $reporters = Double::instance();
 
             expect($reporters)->toReceive('dispatch')->with('start', ['total' => 2])->ordered;
             expect($reporters)->toReceive('dispatch')->with('suiteStart', $describe)->ordered;
@@ -854,7 +854,7 @@ describe("Suite", function() {
 
             });
 
-            $reporters = Stub::create();
+            $reporters = Double::instance();
 
             expect($reporters)->toReceive('dispatch')->with('start', ['total' => 2])->ordered;
             expect($reporters)->toReceive('dispatch')->with('suiteStart', $describe)->ordered;
@@ -1213,7 +1213,7 @@ describe("Suite", function() {
 
             $describe = $this->suite->describe("", function() {});
 
-            $reporters = Stub::create();
+            $reporters = Double::instance();
             $this->suite->run(['reporters' => $reporters]);
 
             expect($this->suite->reporters())->toBe($reporters);
@@ -1228,7 +1228,7 @@ describe("Suite", function() {
 
             $describe = $this->suite->describe("", function() {});
 
-            $reporters = Stub::create();
+            $reporters = Double::instance();
 
             expect($reporters)->toReceive('dispatch')->with('stop', Arg::toBeAnInstanceOf('Kahlan\Summary'));
